Checking Operating Systems
This is still a temporary fix. Until I get an actual server to replace this site version, I'll have to continue adjusting for Safari's form issues. Because of this, I wrote a function that checks for the operating system, and, based off of that information, adjusts the HTML to direct to another form. People wouldn't even know the difference!