textHolder bug fixed
I had to fix a bug that was hitting the text holder id tags. The reason why was that I gave two items with the same id. I may have to fix that later. But for now, as a temporary fix, I gave the paragraph tag an id, deleted the previous code, and went on from there.