Repositioning Google Analytics Tags
For some reason, the data wasn't appearing on Google Analytics. It would do it for a while, but then it would mess up. Because of this, I had to reposition the code that ran that part. Hopefully this fix works. If not, I'll work a bit more on it today.