Keyword Improvements
I generated a whole list of keywords to make sure that we are found by whoever needs violin lessons. I don't know how big GitHub can let me get, but we'll see. I'll check the payment part. If it starts to get expensive, I'll lower the amount of keywords that are being used.