A JavaScript Solution for Detecting Spam
Recently, there has been this particular issue that has been going on with this site: there have been spam concerning a financial robot. After a while, the form feed begins to fill with spam mail in the Submissions tab. I didn't want the spam mail to continue entering the platform, but I didn't want to enable the function either.

Because of this, I wrote a JavaScript function that could detect when a spam message is being written out. By doing so, I am enabling spam detection indirectly.