Data Transfer Issue Fixed
Ok, I think I fixed it now. Turns out that the page was blowing up in one of the lines of code that had to do with grabbing information concerning the instrument that the parent wants to learn. As a fix, I nested all of the conditionals in a conditional statement which only runs if the parent is learning.

Now, I still have to test this. I've tested this on my computer, but I still have to test this on the web.