Removing Operating System Check
There was a JavaScript function that I was using before that ran to check the operating system. The reason why I had this before was in order to find a temporary fix to the Google Forms transfer issue. However, since I have found a more efficient way, I removed the function. There's no need for that function anymore.

I'm going to have to run a few tests to make sure this works, though. I'm going to test the capabilities of every page in this site.