Sitemap + More Changes
I'm right now working on improving the ranking of my website. One of the many things that I did is that I wrote an XML file that has the map of the site. This is mostly to make this work. I also have to work on making all of the code valid as well. Right now, it isn't following the correct standards.