Auto-resizing Video and Updating Sitemap.xml
Basically, I wrote a few more media queries that would help with resizing the video panel. I also updated the sitemap.xml code to show that the index.html page had been updated.

Of course, there's still a lot more to work on with this project, but we're getting there.