Removed Contact Us Disallow
I originally wasn't allowing robots to enter Contact Us. However, I don't have to worry about that anymore. I removed it due to the fact that it wasn't preventing the spam mail from getting through. I've concluded that the spam mail is primarily directed to my form collector. That might be the case.