Updating Keywords
I shortened the amount of keywords that I had before. There were too many keywords earlier, so I decided to shorten them down. Also, due to the fact that the developer contact form is just for the developer, I just replaced the keywords with "Easey Strings Web Developer." That was it.