Updating #textHolder and Adding NoScript Information
For a while, the code for a few of the pages was not valid because I had two items sharing the same id. Sure, they both worked fine, but I wanted to stick with the standards. Because of this, I switched up #textHolder to .textHolder, changing the id to a class.

As for the noscript tag, I wrote it out just in case there are individuals using a browser that has JavaScript turned off. This particular website uses a ton of JavaScript, so some functions may not work if browsers do not have JavaScript enabled.