Updating Colors,, Adding New Pages, and Updating Sitemaps
I was told to add a new page, so I did that. I also changed the media queries a little bit so that they could work with the new page in the list. I also changed the colors so that all of the buttons did not have to have different colors.

I also updated the sitemap since the new page wasn't on the list. This gave me the chance to update the code that automatically updates the XML.

This was some fun coding!