Improving UI
There was an issue in this page where the image would not be directly underneath the list of prices. The fix for this was adding extra spaces before the image. Using that, I was able to forcefully center the image without issues. I made it not draggable in order for it to look as if it was centered by itself.

If I find a better solution, I'll use it to replace this one. However, this was the one that is working for now.