New Section Added
It's quite nice to get back to working on this particular site again. It has been a while since I have worked on it.

I basically added another section to the site which tells the user where our music lessons are being held. I am still trying to decide if I should add a GPS link to that particular area. I might as well talk that over with the director of the music studio.