More Changes
I just added a bit more information to the website concerning where the lessons are being held. I decided that adding a link would not hurt anyone. I added a few images to the site, but I have them hidden for now because I have yet to wait for permission to put pictures of lessons. That may take a while.