Removing REMs
I had several commented pieces of code that I did not need in my site anymore. The plan is to make sure that this site does not have any unnecessary REMs by the end of this year. If they are notes in order to remind another web developer what a particular section of code is for, I will leave it.