Fixed Issue
I had to fix a few things when it came to the sitemap. For some reason, my Python program did not update it correctly, so I had to rewrite the HTML code and then test the program again. I wonder why the program doesn't work like that all the time.