Bringing Back the Developer Contact
I've decided to bring back the developer contact form because I know I won't be able to catch all of the bugs. Therefore, other users can help out with finding them. This is a rather logical conclusion considering the fact that I'm working on other projects at the same time, school included. So, I get quite busy in between.

Of course, I'll just leave it like that for now. I've got to get to the other bit of work that I have to do today.