More OpenGraph Tags
I added a few more OpenGraph meta tags. I also noticed that there was a typo on one of the meta tags, which was probably the reason why it was not popping up on Facebook properly. Well, now it should be fixed. I also inserted locale information. Changed the value of og:type to website as well. Pointed each image to its secure URL.

This has been fun to learn, that's for sure!