Updating Robots.txt
Had to add some more code into that file. For some reason, the reason Facebook was not able to generate the images was because it was not allowed to look through the url. Therefore, I allowed it to have access to all of our website information, including a few of the images. Otherwise, it cannot look through the rest of our images. At least, I hope that is what it does.