More Code Changes
Just discovered that the value of a parameter when it cannot be found is null. Therefore, I extended the logic to include all relevant data except for middleNameValue. That is an optional field, so it does not matter.

Also, I wrote a conditional statement for the third stage that writes in the middle name if and only if it is not null. Otherwise, it writes the other name.