Try-n-Catch Added
Upon trying to see if the redirect code worked, I was surprised to see that it did not redirect. Turns out that an error occurred on former line 447 due to a null value being placed in a function that needed to run a replaceAll operation. Therefore, I surrounded all of the code with a try-catch statement, in which when the error occurs, the redirect happens. It's quite a neat idea.